build 64 bit rcp rlogin rsh

+++ usr/src/cmd/cmd-inet/usr.bin/rcp.c
@@ -1827,7 +1827,7 @@ notzero(char *p, int n)
 {
 	register int result = 0;
 
-	while ((int)p & 3 && --n >= 0)
+	while ((intptr_t)p & 3 && --n >= 0)
 		result |= *p++;
 
 	while ((n -= 4 * sizeof (int)) >= 0) {
